refactor current elasticsearch folder into package
